+++ CHANGELOG.md
@@ -3,10 +3,23 @@
 ## Change History
 All notable changes to the Databricks Labs Data Generator will be documented in this file.
 
-### Version Unreleased
+### Version 0.3.2
 
 #### Changed
+* Adjusted column build phase separation (i.e which select statement is used to build columns) so that a 
+  column with a SQL expression can refer to previously created columns without use of a `baseColumn` attribute
 * Changed build labelling to comply with PEP440
+
+#### Fixed 
+* Fixed compatibility of build with older versions of runtime that rely on `pyparsing` version 2.4.7
+
+#### Added 
+* Parsing of SQL expressions to determine column dependencies
+
+#### Notes
+* The enhancements to build ordering does not change actual order of column building -
+  but adjusts which phase columns are built in 
+
 
 ### Version 0.3.1